backend/auth_service/requirements.txt
aio-pika==9.5.7
aiobotocore==2.24.2
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiormq==6.9.0
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
botocore==1.40.18
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
compose==1.6.2
docker==7.1.0
environs==14.1.1
fastapi==0.115.12
fastapi-cache2==0.2.2
fastapi-limiter==0.1.6
frozenlist==1.7.0
greenlet==3.2.0
h11==0.14.0
hiredis==3.2.1
httptools==0.6.4
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
izulu==0.50.0
Jinja2==3.1.6
jmespath==1.0.1
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
multidict==6.6.4
nats-py==2.10.0
packaging==25.0
pamqp==3.3.0
passlib==1.7.4
pendulum==3.1.0
propcache==0.3.2
psycopg==3.2.6
psycopg-binary==3.2.6
pycron==3.1.2
pydantic[email]==2.11.2
pydantic-settings==2.10.1
pydantic_core==2.33.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
requests==2.32.3
segno==1.6.6
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
taskiq==0.11.17
taskiq-dependencies==1.5.7
taskiq-nats==0.5.1
text-unidecode==1.3
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
wrapt==1.17.3
yarl==1.20.1
zipp==3.21.0
gigachat
urllib3
bcrypt
PyJWT
python-dotenv